+++ packages/celer-devtool/src/cds/display.rs
@@ -14,7 +14,7 @@ impl DevServerDisplay {
     pub fn update(&mut self, port: u16, last_update: &str, bundle_context: &bundle::BundleContext) {
         let mut new_display = String::new();
         new_display.push_str("======== Celer Dev Server ========\n");
-        if port != super::config::DEFAULT_PORT {
+        if port != super::DEFAULT_PORT {
             writeln!(new_display, "Port: {port}").unwrap();
         }
 
@@ -23,7 +23,7 @@ impl DevServerDisplay {
 
         new_display.push_str(&bundle_context.get_error_str_or("No issue found\n"));
         new_display.push('\n');
-        if port != super::config::DEFAULT_PORT {
+        if port != super::DEFAULT_PORT {
             writeln!(new_display, "Open https://celer.itntpiston.app/#/dev/{port} in your browser").unwrap();
         }else{
             new_display.push_str("Open https://celer.itntpiston.app/#/dev in your browser\n");