+++ packages/celer-devtool/src/cio/file.rs
@@ -49,3 +49,13 @@ pub fn write_file(content: &[u8], file_name: &str, out_errors: &mut ErrorState) 
         out_errors.add(file_name, format!("Cannot write to {file_name}: {why}"));
     }
 }
+
+pub fn is_celer_module(file_name: &Path) -> bool {
+    let file_name = file_name.file_name().and_then(|x| x.to_str()).unwrap_or("");
+    file_name.ends_with(".celer") || file_name.ends_with(".yaml")
+}
+
+pub fn is_related_module(file_name: &Path) -> bool {
+    let file_name = file_name.file_name().and_then(|x| x.to_str()).unwrap_or("");
+    file_name.ends_with(".celer") || file_name.ends_with(".yaml") || file_name.ends_with(".png")
+}