+++ packages/celer-web-app/src/app/services/DocumentDev.tsx
@@ -1,5 +1,5 @@
 import { Params } from "react-router-dom";
-import { wasmBundle, wasmBundleFromBase64 } from "data/libs";
+import { SourceObject, wasmBundle, wasmBundleFromBase64 } from "data/libs";
 import { Consumer } from "data/util";
 import { Document, DocumentResponse } from "./types";
 
@@ -14,19 +14,28 @@ class DocumentDev implements Document {
 		}
 		this.enableBase64 = enableBase64;
 	}
-	load(callback: Consumer<DocumentResponse>): void {
+	async load(callback: Consumer<DocumentResponse>): Promise<DocumentResponse> {
 		console.log("Connecting to local ws dev server " + this.port); // eslint-disable-line no-console
 		const newws = new WebSocket("ws://localhost:" + this.port);
 		newws.onerror = (e) => {
 			console.error(e);
-			callback({error: "Cannot connect to the dev server. Make sure the dev server is running and refresh the page to try again"});
+			return {
+				error: "Cannot connect to the dev server. Make sure the dev server is running and refresh the page to try again"
+			};
 		};
 		newws.onmessage = (e) => {
 			const dataObject = JSON.parse(e.data);
 			if (this.enableBase64) {
 				if (dataObject.type) {
 					if (dataObject.type === "base64") {
-						const doc = wasmBundleFromBase64(dataObject.data).bundle; //Discard the errors for now
+						const doc: SourceObject | undefined = wasmBundleFromBase64(dataObject.data);
+						if (!doc) {
+							callback({error: "Failed to read bundle."});
+							return;
+						}
+						if (doc._globalError) {
+							callback({error: doc._globalError});
+						}
 						callback({doc});
 					} else {
 						callback({error: "Unknown data type"});
@@ -44,14 +53,14 @@ class DocumentDev implements Document {
 			callback({ status: "Waiting for data"});
 		};
 		this.ws = newws;
-		callback({ status: "Connecting to dev server" });
+		return { status: "Connecting to dev server" };
 	}
 	release(): void {
 		this.ws?.close();
 	}
 
 	getPath(): string {
-		return "dev/" + this.port;
+		return "dev/" + this.port; // Remove with useExpUseNewRecentPath
 	}
 }