+++ packages/celer-web-app/src/app/services/types.ts
@@ -10,11 +10,12 @@ export type DocumentResponse = {
 };
 
 export interface Document {
-    // Executes any async tasks to get the document
-    // and return the document through the callback
-    load(callback: Consumer<DocumentResponse>): void,
+    // Get the document asynchronously.
+    // Further updates to the document can be send to the callback
+    load(callback: Consumer<DocumentResponse>): Promise<DocumentResponse>,
     // Clean up and destroys the service. Any pending tasks must be cancelled
     release(): void,
-    // Get the document path (the part after #/ in the url, for example, gh/username/repo)
+    // DEPRECATED Get the document path (the part after #/ in the url, for example, gh/username/repo)
+    // Remove with useExpUseNewRecentPath
     getPath(): string,
 }